add hcn-init.service.suse service covering wicked
While hcnmgr -s needs to communicate with NetworkManager and
thus the hcn-init.service has to start after it, in wicked
case, we can prepare the actually needed configuration and
let wicked.service starting after apply it.
This avoids to try to setup a potentially obsolete bonding
first (and maybe even run into timeouts) and then modify it
to the current hcn state in device-tree.