Fix re-set primary bonding slave after LPM
Post migration the new sr_iov is expected to be add to its HNV bonding,
as primary slave. The bonding ifcfg is modified correctly to set new
se_iov as primary, wicked ifreload picked up ifcfg configration correctly
but couldnt able to work with backend wicked nanny to propergate this update
to kernel. So kernel still takes the backup device as the prmary one.

Work around this by replace with wicked up and forcing kernel primary slave
with sysfs interface.

Signed-off-by: Mingming Cao <mmc@linux.vnet.ibm.com>